#632 - rename and reorganize
